+++ edison-jobs/src/main/java/de/otto/edison/jobs/eventbus/LogJobStateChangeListener.java
@@ -1,0 +1,21 @@
+package de.otto.edison.jobs.eventbus;
+
+import de.otto.edison.jobs.eventbus.events.StateChangeEvent;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class LogJobStateChangeListener implements JobStateChangeListener {
+
+    public static final Logger LOG = LoggerFactory.getLogger(LogJobStateChangeListener.class);
+
+    @Override
+    public void consumeStateChange(final StateChangeEvent stateChangeEvent) {
+        LOG.info(
+                "jobType='{}' state changed to '{}' ('{}'): {}",
+                stateChangeEvent.getJobType(),
+                stateChangeEvent.getState(),
+                stateChangeEvent.getJobId(),
+                stateChangeEvent.getMessage());
+    }
+
+}