+++ edison-jobs/src/test/java/de/otto/edison/jobs/eventbus/InMemoryEventRubbishBin.java
@@ -1,6 +1,5 @@
 package de.otto.edison.jobs.eventbus;
 
-import de.otto.edison.jobs.eventbus.events.MessageEvent;
 import de.otto.edison.jobs.eventbus.events.StateChangeEvent;
 import org.springframework.context.event.EventListener;
 
@@ -10,28 +9,16 @@ import java.util.List;
 public class InMemoryEventRubbishBin {
 
     private final List<String> stateChangedEvents = new ArrayList<>();
-    private final List<String> messageEvents = new ArrayList<>();
 
     @EventListener
     public void consumeStateChangedEvent(final StateChangeEvent stateChangeEvent) {
         stateChangedEvents.add(stateChangeEvent.getJobId());
     }
-
-    @EventListener
-    public void consumeMessageEvent(final MessageEvent messageEvent) {
-        messageEvents.add(messageEvent.getJobId());
-    }
-
     public List<String> getStateChangedEvents() {
         return stateChangedEvents;
     }
 
-    public List<String> getMessageEvents() {
-        return messageEvents;
-    }
-
     public void clear() {
         stateChangedEvents.clear();
-        messageEvents.clear();
     }
 }