+++ edison-jobs/src/test/java/de/otto/edison/jobs/eventbus/PersistenceJobStateChangeListenerTest.java
@@ -1,8 +1,6 @@
 package de.otto.edison.jobs.eventbus;
 
 import de.otto.edison.jobs.definition.JobDefinition;
-import de.otto.edison.jobs.domain.JobMessage;
-import de.otto.edison.jobs.eventbus.events.MessageEvent;
 import de.otto.edison.jobs.eventbus.events.StateChangeEvent;
 import de.otto.edison.jobs.service.JobRunnable;
 import de.otto.edison.jobs.service.JobService;
@@ -10,32 +8,32 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 
-import java.time.Instant;
 import java.time.OffsetDateTime;
-import java.time.ZoneId;
-import java.util.Optional;
 
-import static de.otto.edison.jobs.domain.Level.INFO;
-import static de.otto.edison.jobs.eventbus.events.MessageEvent.newMessageEvent;
+import static de.otto.edison.jobs.domain.JobMessage.jobMessage;
+import static de.otto.edison.jobs.domain.Level.ERROR;
 import static de.otto.edison.jobs.eventbus.events.StateChangeEvent.State.*;
 import static de.otto.edison.jobs.eventbus.events.StateChangeEvent.newStateChangeEvent;
+import static java.time.Instant.ofEpochMilli;
+import static java.time.OffsetDateTime.ofInstant;
+import static java.time.ZoneId.systemDefault;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
-public class PersistenceJobEventListenerTest {
+public class PersistenceJobStateChangeListenerTest {
 
-    public static final String JOB_ID = "some/job/id";
-    public static final String JOB_TYPE = "jobType";
+    private static final String JOB_ID = "some/job/id";
+    private static final String JOB_TYPE = "jobType";
 
     @Mock
     private JobService jobServiceMock;
     @Mock
     private JobRunnable jobRunnableMock;
 
-    private PersistenceJobEventListener subject;
+    private PersistenceJobStateChangeListener subject;
 
 
 
@@ -45,7 +43,7 @@ public class PersistenceJobEventListenerTest {
 
         when(jobRunnableMock.getJobDefinition()).thenReturn(mock(JobDefinition.class));
 
-        subject = new PersistenceJobEventListener(jobServiceMock);
+        subject = new PersistenceJobStateChangeListener(jobServiceMock);
     }
 
     @Test
@@ -69,7 +67,7 @@ public class PersistenceJobEventListenerTest {
         when(jobRunnableMock.getJobDefinition()).thenReturn(mockDefinition);
         subject.consumeStateChange(stateChangedEvent(DEAD));
 
-        verify(jobServiceMock).killJob(JOB_ID, JOB_TYPE);
+        verify(jobServiceMock).killJob(JOB_ID);
     }
 
     @Test
@@ -80,34 +78,20 @@ public class PersistenceJobEventListenerTest {
     }
 
     @Test
+    public void shouldPersistFailedEvent() throws Exception {
+        final StateChangeEvent event = stateChangedEvent(FAILED);
+        subject.consumeStateChange(event);
+
+        final OffsetDateTime ts = ofInstant(ofEpochMilli(event.getTimestamp()), systemDefault());
+        verify(jobServiceMock).appendMessage(JOB_ID, jobMessage(ERROR, "", ts));
+    }
+
+    @Test
     public void shouldPersistSkippedEvent() throws Exception {
         subject.consumeStateChange(stateChangedEvent(SKIPPED));
 
         verify(jobServiceMock).markSkipped(JOB_ID);
     }
-
-    @Test
-    public void shouldPersistMessage() throws Exception {
-        MessageEvent messageEvent = newMessageEvent(jobRunnableMock, JOB_ID, INFO, "some message", Optional.empty());
-        OffsetDateTime timestamp = OffsetDateTime.ofInstant(Instant.ofEpochMilli(messageEvent.getTimestamp()), ZoneId.systemDefault());
-
-        subject.consumeMessage(messageEvent);
-
-        verify(jobServiceMock).appendMessage(JOB_ID, JobMessage.jobMessage(INFO, "some message", timestamp));
-    }
-
-    @Test
-    public void shouldNotThrowIfSomethingFailsInDatabase() {
-        MessageEvent messageEvent = newMessageEvent(jobRunnableMock, JOB_ID, INFO, "some message", Optional.empty());
-        OffsetDateTime timestamp = OffsetDateTime.ofInstant(Instant.ofEpochMilli(messageEvent.getTimestamp()), ZoneId.systemDefault());
-        final JobMessage expectedJobMessage = JobMessage.jobMessage(INFO, "some message", timestamp);
-        doThrow(new RuntimeException("Miserable failure")).when(jobServiceMock).appendMessage(JOB_ID, expectedJobMessage);
-
-        subject.consumeMessage(messageEvent);
-
-        verify(jobServiceMock).appendMessage(JOB_ID, expectedJobMessage);
-    }
-
 
     @Test
     public void shouldNotThrowIfStateChangeFailsInDatabase() throws Exception {