+++ examples/example-jobs/src/main/java/de/otto/edison/example/jobs/ExampleMetaJob.java
@@ -2,19 +2,21 @@ package de.otto.edison.example.jobs;
 
 import de.otto.edison.jobs.definition.JobDefinition;
 import de.otto.edison.jobs.domain.MetaJobRunnable;
-import de.otto.edison.jobs.eventbus.JobEventPublisher;
 import de.otto.edison.jobs.repository.JobMetaRepository;
-import de.otto.edison.jobs.service.JobMetaService;
+import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.Optional;
 
 import static de.otto.edison.jobs.definition.DefaultJobDefinition.manuallyTriggerableJobDefinition;
+import static de.otto.edison.jobs.domain.JobMarker.JOB;
+import static org.slf4j.LoggerFactory.getLogger;
 
 @Component
 public class ExampleMetaJob extends MetaJobRunnable {
 
+    private static final Logger LOG = getLogger(ExampleMetaJob.class);
     public static final String JOB_TYPE = "ExampleMetaJob";
 
     @Autowired
@@ -33,13 +35,13 @@ public class ExampleMetaJob extends MetaJobRunnable {
     }
 
     @Override
-    public void execute(JobEventPublisher jobEventPublisher) {
-
+    public boolean execute() {
         int lastEntry = getMetaAsInt("lastEntry", 0);
 
         for (int i = lastEntry+1; i <= lastEntry + 10; i++) {
-            jobEventPublisher.info("Processing Item " + i);
+            LOG.info(JOB, "Processing Item {}", i);
             setMeta("lastEntry", i);
         }
+        return true;
     }
 }