+++ examples/example-jobs/src/main/java/de/otto/edison/example/jobs/FooJob.java
@@ -1,8 +1,9 @@
 package de.otto.edison.example.jobs;
 
 import de.otto.edison.jobs.definition.JobDefinition;
-import de.otto.edison.jobs.eventbus.JobEventPublisher;
+import de.otto.edison.jobs.domain.JobMarker;
 import de.otto.edison.jobs.service.JobRunnable;
+import org.slf4j.Logger;
 import org.springframework.stereotype.Component;
 
 import java.util.Optional;
@@ -11,6 +12,7 @@ import java.util.Random;
 import static de.otto.edison.jobs.definition.DefaultJobDefinition.fixedDelayJobDefinition;
 import static java.lang.Thread.sleep;
 import static java.time.Duration.ofHours;
+import static org.slf4j.LoggerFactory.getLogger;
 
 /**
  * @author Guido Steinacker
@@ -18,6 +20,8 @@ import static java.time.Duration.ofHours;
  */
 @Component
 public class FooJob implements JobRunnable {
+
+    private static final Logger LOG = getLogger(FooJob.class);
 
     @Override
     public JobDefinition getJobDefinition() {
@@ -32,18 +36,19 @@ public class FooJob implements JobRunnable {
     }
 
     @Override
-    public void execute(final JobEventPublisher jobEventPublisher) {
+    public boolean execute() {
         for (int i = 0; i < 60; ++i) {
-            doSomeHardWork(jobEventPublisher);
+            doSomeHardWork();
         }
+        return true;
     }
 
-    private void doSomeHardWork(final JobEventPublisher jobEventPublisher) {
+    private void doSomeHardWork() {
         try {
-            jobEventPublisher.info("Still doing some hard work...");
+            LOG.info(JobMarker.JOB, "Still doing some hard work...");
             sleep(new Random(42).nextInt(2000));
         } catch (final InterruptedException e) {
-            jobEventPublisher.error(e.getMessage());
+            LOG.error(JobMarker.JOB, e.getMessage());
         }
     }
 }